requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
joblib>=1.4.2
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
xgboost>=2.0.0
seaborn>=0.13.0
matplotlib>=3.8.0
openpyxl>=3.1.2
jupyter>=1.0.0
nbconvert>=7.16.0
httpx>=0.27.0

google-genai>=0.3.0
